moved auth result data model to models folder
