+++ Team10_JournalApp/Models/Firebase/AuthDataResultModel.swift
@@ -1,0 +1,21 @@
+//
+//  AuthDataResultModel.swift
+//  Team10_JournalApp
+//
+//  Created by ujwal joshi on 11/19/24.
+//
+
+import Foundation
+import FirebaseAuth
+
+struct AuthDataResultModel {
+    let uid: String
+    let email: String?
+    let photoUrl: String?
+    
+    init(user: User) {
+        self.uid = user.uid
+        self.email = user.email
+        self.photoUrl = user.photoURL?.absoluteString
+    }
+}
