+++ Team10_JournalApp/Utilities/Firebase/AuthenticationManager.swift
@@ -8,17 +8,7 @@
 import Foundation
 import FirebaseAuth
 
-struct AuthDataResultModel {
-    let uid: String
-    let email: String?
-    let photoUrl: String?
-    
-    init(user: User) {
-        self.uid = user.uid
-        self.email = user.email
-        self.photoUrl = user.photoURL?.absoluteString
-    }
-}
+
 
 final class AuthenticationManager {
     static let shared = AuthenticationManager()
